feat: add platform stub for unified TTS
