+++ lib/services/tts_platform_stub.dart
@@ -1,0 +1,10 @@
+class Platform {
+  static const Map<String, String> environment = {};
+}
+
+class SocketException implements Exception {
+  final String message;
+  SocketException([this.message = '']);
+  @override
+  String toString() => 'SocketException: $message';
+}
